backend/pyproject.toml
[tool.poetry]
name = "jean-backend"
version = "0.1.0"
description = "Backend MCP server for JEAN Personal Memory Layer"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
asyncpg = "^0.29.0"
python-dotenv = "^1.0.1"
google-generativeai = "^0.5.4"
pydantic = "^2.7.1"
pydantic-settings = "^2.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 